cur quest update

+++ Assets/Scripts/Core/Application/Requirements/Checkers/SaveState/ReqUnitDeadChecker.cs
@@ -18,7 +18,7 @@ namespace Core.Application.Requirements.Checkers.SaveState
         protected override float GetProgressInternal(ReqUnitDead req)
         {
             var progress = GetEventProgress(req);
-            return Mathf.Clamp01(progress/req.Count);
+            return Mathf.Clamp01((float)progress/(float)req.Count);
         }
 
         protected override bool CheckInternal(ReqUnitDead req)
